golf-geometry-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
# anthropic>=0.44.0  # Optional: only needed for Vision-based tracing fallback
pydantic>=2.9.0
pydantic-settings>=2.6.0
Pillow>=10.4.0
aiosqlite>=0.20.0
cachetools>=5.5.0
python-dotenv>=1.0.0
